LayoutTests/imported/w3c/web-platform-tests/css/css-logical/animations/float-interpolation-expected.txt


PASS CSS Transitions: property <float> from [left] to [right] at (-0.3) should be [right]
PASS CSS Transitions: property <float> from [left] to [right] at (0) should be [right]
PASS CSS Transitions: property <float> from [left] to [right] at (0.3) should be [right]
PASS CSS Transitions: property <float> from [left] to [right] at (0.5) should be [right]
PASS CSS Transitions: property <float> from [left] to [right] at (0.6) should be [right]
PASS CSS Transitions: property <float> from [left] to [right] at (1) should be [right]
PASS CSS Transitions: property <float> from [left] to [right] at (1.5) should be [right]
PASS CSS Transitions with transition: all: property <float> from [left] to [right] at (-0.3) should be [right]
PASS CSS Transitions with transition: all: property <float> from [left] to [right] at (0) should be [right]
PASS CSS Transitions with transition: all: property <float> from [left] to [right] at (0.3) should be [right]
PASS CSS Transitions with transition: all: property <float> from [left] to [right] at (0.5) should be [right]
PASS CSS Transitions with transition: all: property <float> from [left] to [right] at (0.6) should be [right]
PASS CSS Transitions with transition: all: property <float> from [left] to [right] at (1) should be [right]
PASS CSS Transitions with transition: all: property <float> from [left] to [right] at (1.5) should be [right]
FAIL CSS Animations: property <float> from [left] to [right] at (-0.3) should be [left] assert_equals: expected "left " but got "none "
FAIL CSS Animations: property <float> from [left] to [right] at (0) should be [left] assert_equals: expected "left " but got "none "
FAIL CSS Animations: property <float> from [left] to [right] at (0.3) should be [left] assert_equals: expected "left " but got "none "
FAIL CSS Animations: property <float> from [left] to [right] at (0.5) should be [right] assert_equals: expected "right " but got "none "
FAIL CSS Animations: property <float> from [left] to [right] at (0.6) should be [right] assert_equals: expected "right " but got "none "
FAIL CSS Animations: property <float> from [left] to [right] at (1) should be [right] assert_equals: expected "right " but got "none "
FAIL CSS Animations: property <float> from [left] to [right] at (1.5) should be [right] assert_equals: expected "right " but got "none "
FAIL Web Animations: property <float> from [left] to [right] at (-0.3) should be [left] assert_equals: expected "left " but got "none "
FAIL Web Animations: property <float> from [left] to [right] at (0) should be [left] assert_equals: expected "left " but got "none "
FAIL Web Animations: property <float> from [left] to [right] at (0.3) should be [left] assert_equals: expected "left " but got "none "
FAIL Web Animations: property <float> from [left] to [right] at (0.5) should be [right] assert_equals: expected "right " but got "none "
FAIL Web Animations: property <float> from [left] to [right] at (0.6) should be [right] assert_equals: expected "right " but got "none "
FAIL Web Animations: property <float> from [left] to [right] at (1) should be [right] assert_equals: expected "right " but got "none "
FAIL Web Animations: property <float> from [left] to [right] at (1.5) should be [right] assert_equals: expected "right " but got "none "
